third_party/WebKit/LayoutTests/inspector/sources/debugger/show-function-definition.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/console-test.js"></script>
<script>

function jumpToMe()
{
    var result = 12345;
    return window.foo || result;
}

function test()
{
    var panel = WebInspector.inspectorView.showPanel("sources");

    InspectorTest.runTestSuite([
        function testRevealFunctionDefinition(next)
        {
            InspectorTest.addSniffer(panel, "showUISourceCode", showUISourceCodeHook);
            WebInspector.context.flavor(WebInspector.ExecutionContext).evaluate("jumpToMe", "", false, true, false, false, didGetFunction);

            function didGetFunction(funcObject, wasThrown)
            {
                var error = !funcObject || wasThrown;
                InspectorTest.assertTrue(!error);
                panel._showFunctionDefinition(funcObject);
            }

            function showUISourceCodeHook(uiSourceCode, lineNumber, columnNumber, forceShowInPanel)
            {
                // lineNumber and columnNumber are 0-based
                ++lineNumber;
                ++columnNumber;
                InspectorTest.addResult("Function location revealed: [" + lineNumber + ":" + columnNumber + "]");
                next();
            }
        },

        function testDumpFunctionDefinition(next)
        {
            InspectorTest.waitUntilNthMessageReceived(2, onConsoleMessagesReceived);
            WebInspector.ConsoleModel.evaluateCommandInConsole(WebInspector.context.flavor(WebInspector.ExecutionContext), "jumpToMe");

            function onConsoleMessagesReceived()
            {
                var messages = [];
                InspectorTest.disableConsoleViewport();
                var viewMessages = WebInspector.ConsolePanel._view()._visibleViewMessages;
                for (var i = 0; i < viewMessages.length; ++i) {
                    var uiMessage = viewMessages[i];
                    var element = uiMessage.contentElement();
                    messages.push(element.textContent);
                }

                InspectorTest.addResult(messages.join("\n"));

                next();
            }
        }
    ]);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that "Show Function Definition" jumps to the correct location.
</p>

</body>
</html>
